@@ -4,58 +4,65 @@
   "license": "MIT",
   "scripts": {
     "ng": "nx",
-    "postinstall": "node ./decorate-angular-cli.js && ngcc --properties es2015 browser module main",
+    "postinstall": "node ./decorate-angular-cli.js && ngcc --properties es2020 browser module main",
     "start": "nx serve",
     "build": "nx build",
-    "test": "nx test"
+    "test": "nx test",
+    "prepare": "husky install"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.3.0",
-    "@angular/common": "~13.3.0",
-    "@angular/compiler": "~13.3.0",
-    "@angular/core": "~13.3.0",
-    "@angular/forms": "~13.3.0",
-    "@angular/platform-browser": "~13.3.0",
-    "@angular/platform-browser-dynamic": "~13.3.0",
-    "@angular/router": "~13.3.0",
-    "@nrwl/angular": "13.10.6",
-    "rxjs": "~7.4.0",
+    "@angular/animations": "14.2.12",
+    "@angular/common": "14.2.12",
+    "@angular/compiler": "14.2.12",
+    "@angular/core": "14.2.12",
+    "@angular/forms": "14.2.12",
+    "@angular/platform-browser": "14.2.12",
+    "@angular/platform-browser-dynamic": "14.2.12",
+    "@angular/router": "14.2.12",
+    "@nrwl/angular": "14.8.6",
+    "rxjs": "~7.5.0",
     "tslib": "^2.0.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.3.0",
-    "@angular-eslint/eslint-plugin": "~13.1.0",
-    "@angular-eslint/eslint-plugin-template": "~13.1.0",
-    "@angular-eslint/template-parser": "~13.1.0",
-    "@angular/cli": "~13.3.0",
-    "@angular/compiler-cli": "~13.3.0",
-    "@angular/language-service": "~13.3.0",
-    "@nrwl/cli": "13.10.6",
-    "@nrwl/cypress": "13.10.6",
-    "@nrwl/eslint-plugin-nx": "13.10.6",
-    "@nrwl/jest": "13.10.6",
-    "@nrwl/linter": "13.10.6",
-    "@nrwl/workspace": "13.10.6",
-    "@types/jest": "27.0.2",
+    "@angular-devkit/build-angular": "14.2.10",
+    "@angular-eslint/eslint-plugin": "14.0.4",
+    "@angular-eslint/eslint-plugin-template": "14.0.4",
+    "@angular-eslint/template-parser": "14.0.4",
+    "@angular/cli": "~14.2.0",
+    "@angular/compiler-cli": "14.2.12",
+    "@angular/language-service": "14.2.12",
+    "@commitlint/cli": "^17.0.0",
+    "@commitlint/config-angular": "^17.0.0",
+    "@jscutlery/semver": "^2.29.3",
+    "@nrwl/cli": "14.8.6",
+    "@nrwl/cypress": "14.8.6",
+    "@nrwl/eslint-plugin-nx": "14.8.6",
+    "@nrwl/jest": "14.8.6",
+    "@nrwl/linter": "14.8.6",
+    "@nrwl/workspace": "14.8.6",
+    "@types/jest": "28.1.8",
     "@types/node": "16.11.7",
-    "@typescript-eslint/eslint-plugin": "~5.18.0",
-    "@typescript-eslint/parser": "~5.18.0",
+    "@typescript-eslint/eslint-plugin": "5.48.2",
+    "@typescript-eslint/parser": "5.48.2",
     "cypress": "^9.1.0",
-    "eslint": "~8.12.0",
+    "eslint": "8.15.0",
     "eslint-config-prettier": "8.1.0",
     "eslint-plugin-cypress": "^2.10.3",
-    "jest": "27.2.3",
-    "jest-preset-angular": "11.1.1",
-    "ng-packagr": "^13.0.0",
-    "nx": "13.10.6",
+    "husky": "^8.0.0",
+    "jest": "28.1.3",
+    "jest-environment-jsdom": "28.1.1",
+    "jest-preset-angular": "12.2.5",
+    "ng-packagr": "14.2.2",
+    "nx": "14.8.6",
     "postcss": "^8.4.5",
-    "postcss-import": "^14.0.2",
-    "postcss-preset-env": "^6.7.0",
-    "postcss-url": "^10.1.1",
-    "prettier": "^2.5.1",
-    "ts-jest": "27.0.5",
-    "typescript": "~4.6.2"
+    "postcss-import": "14.1.0",
+    "postcss-preset-env": "7.5.0",
+    "postcss-url": "10.1.3",
+    "prettier": "2.8.3",
+    "ts-jest": "28.0.8",
+    "ts-node": "10.9.1",
+    "typescript": "4.8.4"
   }
 }
